提升性能

+++ carrying-server/src/main/java/com/github/ompc/carrying/server/provider/CarryingProvider.java
@@ -82,7 +82,10 @@ public class CarryingProvider {
                                             synchronized (dos) {
                                                 dos.writeInt(response.getSequence());
                                                 dos.writeInt(response.getLineNumber());
-                                                dos.writeInt(response.getDataLength());
+
+                                                // Hack for LINE.max <= 200B
+                                                dos.writeByte(response.getDataLength());
+
                                                 dos.write(response.getData());
                                             }
                                             dos.flush();
